Move storage code from gaze_producer into global scope
